+++ src/algoritmos/hash/hashSondQuadratica.h
@@ -11,7 +11,7 @@ class HashSondQuad{
         int capacidade;
         int tamanho; 
         NoHash *aux; 
-        unsigned long long int numColisoes;
+        unsigned long numColisoes;
   
     public: 
         HashSondQuad(int capacidade);
@@ -22,6 +22,6 @@ class HashSondQuad{
         bool estaVazia();
         bool estaCheia();
         void display();
-        unsigned long long int getNumColisoes();
+        unsigned long getNumColisoes();
 
 };
