+++ src/algoritmos/ordenacao/GnomeSort.h
@@ -2,6 +2,6 @@
 #define GNOMESORT_H
 
 // Função de ordenação gnome sort
-void gnomeSort(int arr[], int n, unsigned long long int *numComparacoes, unsigned long long int *numTrocas) ;
+void gnomeSort(int arr[], int n, unsigned long *numComparacoes, unsigned long *numTrocas) ;
 
 #endif // GNOMESORT_H
