+++ src/algoritmos/ordenacao/quickSortInsertion.h
@@ -1,3 +1,3 @@
 #include <fstream>
 
-void insertionQuickSort(int *a, int inicio, int fim, int m, unsigned long long int *numComparacoes, unsigned long long int *numTrocas);
+void insertionQuickSort(int *a, int inicio, int fim, int m, unsigned long *numComparacoes, unsigned long *numTrocas);
